repr and str for qi.Timer.
